Improved ParaView interacivity
Fixed a bug that would cause ParaView models to rotate when interacting
with the context menu. Partially addressed a bug with combining
simultaneous rotate and zoom actions. ParaView now applies a large
erroneous rotation, followed entirely by normal zooming until the mouse
is released and clicked again. This is an improvement over the original
behavior, which could cause ParaView to zoom very far in or out. 

Signed-off-by: Robert Smith <smithrw@ornl.gov>