Fixed error with connections not renewing
A subtle bug caused Plot Editors to continue displaying a message that a
connection with a given name was not configured when it was removed from
the preferences table, saved, then re-added. The Preferences menu will
now reuse the connection with the old name, so that resources using it
will be aware of it being reconnected.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>