Added persistable geometry representation
Created the PersistableShape class and subclasses. These classes
represent the ShapeMeshes in a CSG tree for the Geometry Editor, pared
down to the minimal amount of information necessary and annotated to be
persistable through JAXB in a very compact way that is agnostic to the
VizService used to visualize the objects.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>