Fixed Mesh Editor bugs
Fixed a bug where selecting a vertex through the Mesh Element View would
cause the Mesh Element View's trees to collapse.

Fixed a bug where selecting a polygon through the Mesh Element View
while the Mesh Editor was in Add Mode would create an extraneous copy of
the polygon to be listed in the Mesh Element View.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>