Jay/model (#4)
* Added transparency controls to Geometry Editor

The Geometry Editor now fully supports drawing objects in either
wireframe or transparent modes. This is handled through the
Transformation View, which now has a combo box for the selected shape,
allowing the user to switch individual objects between being rendered as
solid, wireframe, or fully transparent.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Updated tests

Updated some tests for transparency and wireframe capabilities. 

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Further work on tests

Did more work on the tests for the ITransparent* and IWireframe*
classes.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Added a rudimentary geometry EMF model.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>

* Expanded model

Expanded the data model to include a view as well as an INode so that
the Shape can be used with Operators in a CSG tree.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Touched up Geometry model.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>

* Expanded geometry model

Added observer pattern and shape properties to the model.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Expanded ShapeChangeEvent

Replaced the ShapeChangeEvent enumeration with a ShapeChangeEvent class
which has a ShapeChangeEventType (the old enumeration) and a string
value, to provide greater detail about the change. 

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Removed subclass-specific propery getters and setters

Removed the getters and setters for attributes like a sphere's radius,
so that all changes to a shape's properties must be sent through the
setProperty() function.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Fixed Plot Editor Icons

The ParaView and VisIt Plot Editors now display their button images
correctly on Windows.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Added Mesh Editor opacity controls

The Properties view for parts in the Mesh Editor now has a box allowing
the user to switch individual pieces between being displayed in solid,
wireframe, or transparent display types inside the editor. 

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Fixed Mesh Editor bugs

Fixed a bug where selecting a vertex through the Mesh Element View would
cause the Mesh Element View's trees to collapse.

Fixed a bug where selecting a polygon through the Mesh Element View
while the Mesh Editor was in Add Mode would create an extraneous copy of
the polygon to be listed in the Mesh Element View.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Added IGeometryImporter and ASCII-based STL implementation stub.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>

* Refined model to better handle resources and eliminate extra references.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>

* Fixed VisIt connection bugs on Linux

Fixed a bug where the Visit Java Client would be unable to properly
locate the VisIt executable inside the file directory and another where
Linux would sometimes be unable to correctly detect whether a directory
existed when searching for VisIt.

Signed-off-by: Robert Smith <SmithRW@ornl.gov>

* Fixed target platform.

Signed-off-by: Jay Jay Billings <billingsjj@ornl.gov>